Write project deletion code and update code
Wrote frontend functions to delete the seletected project from the left
panel list of projects and connected the backend function to the
corresponding event listener. Used the array find method to find the
correct backend project matching the frontend project for deletion.
Repurposed the closeNewProjectForm function to work with other forms
like the deletion form.