apply correct tasks and dates in the frontend
Removed all tasks from the frontend when user closes the project to
prevent previous tasks in the frontend displaying on other projects.
Created applyExistingTasks function to create tasks in the frontend
based on backend information when user selects a project. Corrected the
date format in convertCalendarDate to be consistent with convertDate.