implement task deletion
Made deletion form, used array find method to find the task to delete
and used the array find index method to find the task's index and then
applied both into the deleteBackendTask fuction. Used a while loop to
remove all tasks from the task area, and used applyExistingTasks to
show only the tasks that hadn't been deleted. Renamed the
fillProjectName function as fillItemName and moved it to the
DOM_basic_functions module since I used it with tasks as well.